start of admin.barrier api rework
