+++ packages/web-api/test/types/methods/admin.barriers.test-d.ts
@@ -1,0 +1,11 @@
+import { expectAssignable, expectError } from 'tsd';
+import { WebClient } from '../../../src/WebClient';
+
+const web = new WebClient('TOKEN');
+
+// admin.barriers.create
+// -- sad path
+expectError(web.admin.barriers.create()); // lacking argument
+expectError(web.admin.barriers.create({})); // empty argument
+// -- happy path
+expectAssignable<Parameters<typeof web.admin.barriers.create>>([{}]); // all optional args
